Sensitivity: move the episode_results reader out of dataset.py (review)
- Move from_episode_results and its helpers to a new episode_results_reader.py module, exposed
  as dataset_from_episode_results(); dataset.py is now a pure in-memory container that no longer
  knows the recorder's on-disk format.
- Break the reader into small composable helpers (_read_rows, _discover_factor_values,
  _build_factor_columns, _warn_if_unevenly_sampled, _build_outcome_columns); the entry point is
  a short composition that returns a SensitivityDataset directly.
- Update generate_report, the analyzer error message, and the tests to the new entry point.
